Create a top-level package for SNS change events
`event-bridge/stream` & `event-bridge/publish` share some common code
for defining change event messages, and a convenience wrapper for
publishing them to SNS.

The idea of a message to signal change to documents (currently defined
by the triplet: title, server name, and revision) seems to be
something broadly useful, so this changeset creates a top-level
package with code similar to what is already in use.

A subsequent refactor will remove the `event-bridge` specific code in
favor of this.