Upload schema.org JSON-LD to S3 for new HTML
Implements a Lambda that is subscribed to events indicating new HTML
was uploaded to `incoming/`.  MediaWiki is then used to determine the
pages `wikibase_item`, and the Wikidata Query Service to create a
corresponding [schema.org/Thing](https://schema.org/Thing) (where
possible), serialized as JSON-LD.  The result is uploaded to S3 (the
`schema.org/` folder).