Lambda that augments HTML with linked data
Retrieves HTML & linked data from S3, adds the linked data to the HTML
(as JSON-LD), and uploads the result to S3.

Also includes an alternative to the faux-constants approach of the
`env` package used for configuration: A make snippet is used to define
resources.  The snippet is includ-ed in Makefiles used to compile, and
configuration vars are built into the resulting binary ala `-ldflags`.