Combine page queries (#59)
Combines `page` & `pageByName` into a single query (`page`) that
accepts arguments for either an ID, or a name (currently the only two
relevant ways of querying for a page).

Since page names are only unique to a wiki, the `name` argument is a
`PageNameInput` type, with required attributes of authority (wiki
hostname) and the page's name.