Improve validation failures error messages (#56)
Error messages seen in the wild would seem to indicate there are node
names/section titles within a single document that conflict with one
another.  The validation errors propogated up from the AWS client make
it impossible to see which documents are to blame, so this changeset
detects this condition and returns an error with a formatted copy of
the offending Node.