Resolve all wikidata keywords with their labels
Send a request (or series of requests, based on the 50-item limit)
to the wikidata API to resolve all keywords in displayed sections.
Cache the results so we only ask for keywords that are missing the
labels on new views.

Send that information down the component chain so we can use it
to display the human-readable label.